build: Add ability to generate package-level go protos
Currently, this repository only generates cc and py protos. This change adds the ability to generate *.pb.go and *.pb.validate.go files for each package.
- Example: `bazel build //udpa/data/orca/v1:pkg_go`

Note that the option `go_package` is no longer needed in the proto files (and should not be added!). The go_package will be set automatically to match the path.
- Example: udpa_service_orca_v1

Signed-off-by: Teju Nareddy <nareddyt@google.com>