api: added versioning annotations.
It makes sense to put the annotations that we're adding for the Envoy
xDS stable versioning work here, since UDPA will follow a similar
structure.

Also had to upgrade the api_build_system.bzl, to reflect what is
happening Envoy-side, to allow this to play well with Envoy. It is
largely a copy+paste (and in a few plaes stubbing) of what exists
Envoy-side.

Signed-off-by: Harvey Tuch <htuch@google.com>