annotations: add an active package annotation.
In support of https://github.com/envoyproxy/envoy/issues/10355, it would
be helpful for the protoxform aspect to be able to distinguish the
package version that is currently open for active development.

For some packages, this is likely to move to v3 shortly, but due to the
DAG package version structure, this won't be universal.

Once this lands, I will manually tag the v3 derived active packages in
preparation for https://github.com/envoyproxy/envoy/issues/10355.

Signed-off-by: Harvey Tuch <htuch@google.com>